+++ src/distmat-loops/distmat-loops.h
@@ -9,7 +9,7 @@ RcppExport SEXP distmat_loop(SEXP D, SEXP X, SEXP Y,
                              SEXP DIST, SEXP DIST_ARGS,
                              SEXP FILL_TYPE, SEXP MAT_TYPE, SEXP ENDPOINTS);
 
-RcppExport SEXP dtw_lb(SEXP X, SEXP Y, SEXP D, SEXP MARGIN, SEXP DOTS);
+RcppExport SEXP dtw_lb(SEXP X, SEXP Y, SEXP D, SEXP MARGIN, SEXP DOTS, SEXP NUM_THREADS);
 
 } // namespace dtwclust